Add client ids & more message context
We were debugging an issue and so needed ids to identify messages from
different clients. We also switch to the 'debug' library for nicer
logging.

Finally we pass in the 'cwd' & args to the compile message so that we
have something to compare messages with to find out of any are redundant
due to no file changes between compiles.