Adjust logging
So that we only enable the verbose server logs if there is an env var
set. I've tried making it a command line flag but then the flag parsing
gets confused with any flags & arguments going to elm-make itself. Maybe
when we split the server off to a separate file.