Add priority concept
So that lower priority jobs are killed to make way for higher priority
jobs. The killed jobs are added back on to the queue so that they are
picked up and processed. The clients of the killed jobs do not know that
they have been killed so they hang around until the job is finally
processed.