Remove file watching
I hoped that we'd be able to use it somehow in order to figure out if we
needed to rebuild or could just send the results from the last build
back through.

It still might be an option but the compile will already be quicker on
the second time around so not much need to put loads of effort in from
our side.